% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COG_stroke.r
\name{COG_stroke}
\alias{COG_stroke}
\title{Compute Stroke-Based Center of Gravity (COG)}
\usage{
COG_stroke(img, origin = c("bottomleft", "topleft"))
}
\arguments{
\item{img}{An image input, either a file path to an image file (e.g., PNG, JPEG),
or a \code{cimg} object from the \pkg{imager} package. The image should be in binary form,
with foreground (glyph) values not equal to 1 and background values equal to 1.}

\item{origin}{A character string indicating the location of the image origin.
Use \code{"bottomleft"} (default) if the y-axis increases upward (Cartesian),
or \code{"topleft"} if the y-axis increases downward (as in image arrays).}
}
\value{
A list containing:
\describe{
\item{\code{statistics}}{A data frame with computed center coordinates (original, trimmed, and standardized),
margins around the glyph, and original image dimensions. The center is calculated based on the stroke region}
\item{\code{strokes}}{A data frame of (x, y) coordinates representing the stroke (non-white) region.}
}
}
\description{
Computes the center of gravity (COG) of a character-like binary image using its stroke region.
The COG is defined as the mean of the (x, y) coordinates of all non-white pixels.
}
