% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img-data.r
\docType{data}
\name{img_A}
\alias{img_A}
\title{Example Image A}
\format{
A \code{\link[imager:cimg]{cimg}} object of size 500 × 500 × 1 × 1 (width × height × depth × spectrum).
}
\source{
Created for illustration in the coglyphr package.
}
\usage{
data(img_A)
}
\description{
A binary example image representing the letter "A".
This image is used for demonstrating center-of-gravity calculations in the coglyphr package.
}
\keyword{datasets}
